+++ packages/twenty-server/src/engine/metadata-modules/logic-function/types/update-logic-function-metadata-params.type.ts
@@ -1,4 +1,4 @@
-import {
+import type {
   CronTriggerSettings,
   DatabaseEventTriggerSettings,
   HttpRouteTriggerSettings,
