+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-version-step/__tests__/workflow-version-step-operations.workspace-service.spec.ts
@@ -2,14 +2,15 @@ import { Test, type TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
 
 import { AiAgentRoleService } from 'src/engine/metadata-modules/ai/ai-agent-role/ai-agent-role.service';
-import { AgentService } from 'src/engine/metadata-modules/ai/ai-agent/agent.service';
+import { AgentEntity } from 'src/engine/metadata-modules/ai/ai-agent/entities/agent.entity';
 import { createEmptyAllFlatEntityMaps } from 'src/engine/metadata-modules/flat-entity/constant/create-empty-all-flat-entity-maps.constant';
 import { WorkspaceManyOrAllFlatEntityMapsCacheService } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.service';
 import { LogicFunctionRuntime } from 'src/engine/metadata-modules/logic-function/logic-function.entity';
-import { LogicFunctionMetadataService } from 'src/engine/metadata-modules/logic-function/services/logic-function-metadata.service';
+import { LogicFunctionService } from 'src/engine/metadata-modules/logic-function/services/logic-function.service';
 import { type FlatLogicFunction } from 'src/engine/metadata-modules/logic-function/types/flat-logic-function.type';
 import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadata/object-metadata.entity';
 import { RoleTargetEntity } from 'src/engine/metadata-modules/role-target/role-target.entity';
+import { RoleEntity } from 'src/engine/metadata-modules/role/role.entity';
 import { GlobalWorkspaceOrmManager } from 'src/engine/twenty-orm/global-workspace-datasource/global-workspace-orm.manager';
 import { WorkspaceCacheService } from 'src/engine/workspace-cache/services/workspace-cache.service';
 import { WorkflowCommonWorkspaceService } from 'src/modules/workflow/common/workspace-services/workflow-common.workspace-service';
@@ -25,10 +26,11 @@ const mockWorkspaceId = 'workspace-id';
 describe('WorkflowVersionStepOperationsWorkspaceService', () => {
   let service: WorkflowVersionStepOperationsWorkspaceService;
   let globalWorkspaceOrmManager: jest.Mocked<GlobalWorkspaceOrmManager>;
-  let logicFunctionMetadataService: jest.Mocked<LogicFunctionMetadataService>;
+  let logicFunctionService: jest.Mocked<LogicFunctionService>;
   let codeStepBuildService: jest.Mocked<CodeStepBuildService>;
-  let agentService: jest.Mocked<AgentService>;
+  let agentRepository: jest.Mocked<any>;
   let roleTargetRepository: jest.Mocked<any>;
+  let roleRepository: jest.Mocked<any>;
   let objectMetadataRepository: jest.Mocked<any>;
   let workflowCommonWorkspaceService: jest.Mocked<WorkflowCommonWorkspaceService>;
   let aiAgentRoleService: jest.Mocked<AiAgentRoleService>;
@@ -89,20 +91,24 @@ describe('WorkflowVersionStepOperationsWorkspaceService', () => {
       }),
     } as unknown as jest.Mocked<CodeStepBuildService>;
 
-    logicFunctionMetadataService = {
-      createOne: jest.fn(),
-      destroyOne: jest.fn(),
-    } as unknown as jest.Mocked<LogicFunctionMetadataService>;
+    logicFunctionService = {
+      createOneFromParams: jest.fn(),
+      deleteOne: jest.fn(),
+    } as unknown as jest.Mocked<LogicFunctionService>;
 
-    agentService = {
-      deleteManyAgents: jest.fn().mockResolvedValue([]),
-      findOneAgentById: jest.fn(),
-      createOneAgent: jest.fn(),
-    } as unknown as jest.Mocked<AgentService>;
+    agentRepository = {
+      findOne: jest.fn(),
+      delete: jest.fn(),
+    };
 
     roleTargetRepository = {
       findOne: jest.fn(),
       count: jest.fn(),
+    };
+
+    roleRepository = {
+      findOne: jest.fn(),
+      delete: jest.fn(),
     };
 
     objectMetadataRepository = {
@@ -132,20 +138,24 @@ describe('WorkflowVersionStepOperationsWorkspaceService', () => {
           useValue: globalWorkspaceOrmManager,
         },
         {
-          provide: LogicFunctionMetadataService,
-          useValue: logicFunctionMetadataService,
+          provide: LogicFunctionService,
+          useValue: logicFunctionService,
         },
         {
           provide: CodeStepBuildService,
           useValue: codeStepBuildService,
         },
         {
-          provide: AgentService,
-          useValue: agentService,
+          provide: getRepositoryToken(AgentEntity),
+          useValue: agentRepository,
         },
         {
           provide: getRepositoryToken(RoleTargetEntity),
           useValue: roleTargetRepository,
+        },
+        {
+          provide: getRepositoryToken(RoleEntity),
+          useValue: roleRepository,
         },
         {
           provide: getRepositoryToken(ObjectMetadataEntity),
@@ -203,7 +213,7 @@ describe('WorkflowVersionStepOperationsWorkspaceService', () => {
         workspaceId: mockWorkspaceId,
       });
 
-      expect(logicFunctionMetadataService.destroyOne).toHaveBeenCalledWith({
+      expect(logicFunctionService.deleteOne).toHaveBeenCalledWith({
         id: 'function-id',
         workspaceId: mockWorkspaceId,
       });
@@ -229,13 +239,15 @@ describe('WorkflowVersionStepOperationsWorkspaceService', () => {
         },
       } as unknown as WorkflowAction;
 
+      agentRepository.findOne.mockResolvedValue({ id: 'agent-id' });
+
       await service.runWorkflowVersionStepDeletionSideEffects({
         step,
         workspaceId: mockWorkspaceId,
       });
 
-      expect(agentService.deleteManyAgents).toHaveBeenCalledWith({
-        ids: ['agent-id'],
+      expect(agentRepository.delete).toHaveBeenCalledWith({
+        id: 'agent-id',
         workspaceId: mockWorkspaceId,
       });
     });
@@ -260,6 +272,7 @@ describe('WorkflowVersionStepOperationsWorkspaceService', () => {
         },
       } as unknown as WorkflowAction;
 
+      agentRepository.findOne.mockResolvedValue({ id: 'agent-id' });
       roleTargetRepository.findOne.mockResolvedValue({
         id: 'role-target-id',
         roleId: 'role-id',
@@ -270,8 +283,8 @@ describe('WorkflowVersionStepOperationsWorkspaceService', () => {
         workspaceId: mockWorkspaceId,
       });
 
-      expect(agentService.deleteManyAgents).toHaveBeenCalledWith({
-        ids: ['agent-id'],
+      expect(agentRepository.delete).toHaveBeenCalledWith({
+        id: 'agent-id',
         workspaceId: mockWorkspaceId,
       });
       expect(
@@ -297,21 +310,21 @@ describe('WorkflowVersionStepOperationsWorkspaceService', () => {
         runtime: LogicFunctionRuntime.NODE22,
         timeoutSeconds: 30,
         sourceHandlerPath: 'src/index.ts',
-        builtHandlerPath: 'src/index.mjs',
+        builtHandlerPath: 'index.mjs',
         handlerName: 'main',
         checksum: null,
         toolInputSchema: null,
         isTool: false,
+        isBuildUpToDate: true,
         universalIdentifier: 'universal-id',
         applicationId: 'application-id',
         applicationUniversalIdentifier: 'application-universal-id',
         cronTriggerSettings: null,
         databaseEventTriggerSettings: null,
         httpRouteTriggerSettings: null,
-        isBuildUpToDate: false,
       };
 
-      logicFunctionMetadataService.createOne.mockResolvedValue(
+      logicFunctionService.createOneFromParams.mockResolvedValue(
         mockFlatLogicFunction,
       );
 
@@ -376,13 +389,13 @@ describe('WorkflowVersionStepOperationsWorkspaceService', () => {
         checksum: 'existing-checksum',
         toolInputSchema: null,
         isTool: false,
+        isBuildUpToDate: true,
         universalIdentifier: 'existing-universal-id',
         applicationId: 'application-id',
         applicationUniversalIdentifier: 'application-universal-id',
         cronTriggerSettings: null,
         databaseEventTriggerSettings: null,
         httpRouteTriggerSettings: null,
-        isBuildUpToDate: true,
       };
 
       const mockNewFlatLogicFunction: FlatLogicFunction = {
@@ -396,18 +409,18 @@ describe('WorkflowVersionStepOperationsWorkspaceService', () => {
         runtime: LogicFunctionRuntime.NODE22,
         timeoutSeconds: 30,
         sourceHandlerPath: 'src/index.ts',
-        builtHandlerPath: 'src/index.mjs',
+        builtHandlerPath: 'index.mjs',
         handlerName: 'main',
         checksum: null,
         toolInputSchema: null,
         isTool: false,
+        isBuildUpToDate: true,
         universalIdentifier: 'universal-id',
         applicationId: 'application-id',
         applicationUniversalIdentifier: 'application-universal-id',
         cronTriggerSettings: null,
         databaseEventTriggerSettings: null,
         httpRouteTriggerSettings: null,
-        isBuildUpToDate: false,
       };
 
       const emptyMaps = createEmptyAllFlatEntityMaps();
@@ -428,7 +441,7 @@ describe('WorkflowVersionStepOperationsWorkspaceService', () => {
         },
       );
 
-      logicFunctionMetadataService.createOne.mockResolvedValue(
+      logicFunctionService.createOneFromParams.mockResolvedValue(
         mockNewFlatLogicFunction,
       );
