+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-version-step/workflow-version-step-operations.workspace-service.ts
@@ -9,7 +9,6 @@ import {
 import { isDefined, isValidUuid } from 'twenty-shared/utils';
 import {
   IF_ELSE_BRANCH_POSITION_OFFSETS,
-  getFunctionInputFromInputSchema,
   type StepIfElseBranch,
 } from 'twenty-shared/workflow';
 import { Repository } from 'typeorm';
@@ -18,10 +17,10 @@ import { v4 } from 'uuid';
 import { getFlatFieldsFromFlatObjectMetadata } from 'src/engine/api/graphql/workspace-schema-builder/utils/get-flat-fields-for-flat-object-metadata.util';
 import { type WorkflowStepPositionInput } from 'src/engine/core-modules/workflow/dtos/update-workflow-step-position-input.dto';
 import { AiAgentRoleService } from 'src/engine/metadata-modules/ai/ai-agent-role/ai-agent-role.service';
-import { AgentService } from 'src/engine/metadata-modules/ai/ai-agent/agent.service';
+import { AgentEntity } from 'src/engine/metadata-modules/ai/ai-agent/entities/agent.entity';
 import { DEFAULT_SMART_MODEL } from 'src/engine/metadata-modules/ai/ai-models/constants/ai-models.const';
 import { WorkspaceManyOrAllFlatEntityMapsCacheService } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.service';
-import { LogicFunctionMetadataService } from 'src/engine/metadata-modules/logic-function/services/logic-function-metadata.service';
+import { LogicFunctionService } from 'src/engine/metadata-modules/logic-function/services/logic-function.service';
 import { findFlatLogicFunctionOrThrow } from 'src/engine/metadata-modules/logic-function/utils/find-flat-logic-function-or-throw.util';
 import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadata/object-metadata.entity';
 import { RoleTargetEntity } from 'src/engine/metadata-modules/role-target/role-target.entity';
@@ -65,9 +64,10 @@ const ITERATOR_EMPTY_STEP_POSITION_OFFSET = {
 export class WorkflowVersionStepOperationsWorkspaceService {
   constructor(
     private readonly globalWorkspaceOrmManager: GlobalWorkspaceOrmManager,
-    private readonly logicFunctionMetadataService: LogicFunctionMetadataService,
+    private readonly logicFunctionService: LogicFunctionService,
     private readonly codeStepBuildService: CodeStepBuildService,
-    private readonly agentService: AgentService,
+    @InjectRepository(AgentEntity)
+    private readonly agentRepository: Repository<AgentEntity>,
     @InjectRepository(RoleTargetEntity)
     private readonly roleTargetRepository: Repository<RoleTargetEntity>,
     @InjectRepository(ObjectMetadataEntity)
@@ -87,7 +87,7 @@ export class WorkflowVersionStepOperationsWorkspaceService {
   }) {
     switch (step.type) {
       case WorkflowActionType.CODE: {
-        await this.logicFunctionMetadataService.destroyOne({
+        await this.logicFunctionService.deleteOne({
           id: step.settings.input.logicFunctionId,
           workspaceId,
         });
@@ -98,24 +98,27 @@ export class WorkflowVersionStepOperationsWorkspaceService {
           break;
         }
 
-        const roleTarget = await this.roleTargetRepository.findOne({
-          where: {
-            agentId: step.settings.input.agentId,
-            workspaceId,
-          },
+        const agent = await this.agentRepository.findOne({
+          where: { id: step.settings.input.agentId, workspaceId },
         });
 
-        await this.agentService.deleteManyAgents({
-          ids: [step.settings.input.agentId],
-          workspaceId,
-        });
-
-        if (isDefined(roleTarget?.roleId) && isDefined(roleTarget?.id)) {
-          await this.aiAgentRoleService.deleteAgentOnlyRoleIfUnused({
-            roleId: roleTarget.roleId,
-            roleTargetId: roleTarget.id,
-            workspaceId,
+        if (isDefined(agent)) {
+          const roleTarget = await this.roleTargetRepository.findOne({
+            where: {
+              agentId: agent.id,
+              workspaceId,
+            },
           });
+
+          await this.agentRepository.delete({ id: agent.id, workspaceId });
+
+          if (isDefined(roleTarget?.roleId) && isDefined(roleTarget?.id)) {
+            await this.aiAgentRoleService.deleteAgentOnlyRoleIfUnused({
+              roleId: roleTarget.roleId,
+              roleTargetId: roleTarget.id,
+              workspaceId,
+            });
+          }
         }
         break;
       }
@@ -182,11 +185,7 @@ export class WorkflowVersionStepOperationsWorkspaceService {
               },
               input: {
                 logicFunctionId: newLogicFunction.id,
-                logicFunctionInput: isDefined(newLogicFunction.toolInputSchema)
-                  ? (getFunctionInputFromInputSchema([
-                      newLogicFunction.toolInputSchema,
-                    ])[0] ?? {})
-                  : {},
+                logicFunctionInput: newLogicFunction.toolInputSchema ?? {},
               },
             },
           },
@@ -238,28 +237,6 @@ export class WorkflowVersionStepOperationsWorkspaceService {
             ...baseStep,
             name: 'Send Email',
             type: WorkflowActionType.SEND_EMAIL,
-            settings: {
-              ...BASE_STEP_DEFINITION,
-              input: {
-                connectedAccountId: '',
-                recipients: {
-                  to: '',
-                  cc: '',
-                  bcc: '',
-                },
-                subject: '',
-                body: '',
-              },
-            },
-          },
-        };
-      }
-      case WorkflowActionType.DRAFT_EMAIL: {
-        return {
-          builtStep: {
-            ...baseStep,
-            name: 'Draft Email',
-            type: WorkflowActionType.DRAFT_EMAIL,
             settings: {
               ...BASE_STEP_DEFINITION,
               input: {
@@ -438,20 +415,27 @@ export class WorkflowVersionStepOperationsWorkspaceService {
             workspaceId,
           });
 
-        const newAgent = await this.agentService.createOneAgent(
-          {
-            label:
-              'Workflow Agent' + workflowVersion.workflowId.substring(0, 4),
-            icon: 'IconRobot',
-            description: '',
-            prompt:
-              'You are a helpful AI assistant. Complete the task based on the workflow context.',
-            modelId: DEFAULT_SMART_MODEL,
-            responseFormat: { type: 'text' },
-            isCustom: true,
-          },
+        const newAgent = await this.agentRepository.save({
+          name: 'workflow-service-agent' + v4(),
+          label: 'Workflow Agent' + workflowVersion.workflowId.substring(0, 4),
+          icon: 'IconRobot',
+          description: '',
+          prompt:
+            'You are a helpful AI assistant. Complete the task based on the workflow context.',
+          modelId: DEFAULT_SMART_MODEL,
+          responseFormat: { type: 'text' },
           workspaceId,
-        );
+          isCustom: true,
+        });
+
+        if (!isDefined(newAgent)) {
+          throw new WorkflowVersionStepException(
+            'Failed to create AI Agent step',
+            WorkflowVersionStepExceptionCode.AI_AGENT_STEP_FAILURE,
+          );
+        }
+
+        await this.workspaceCacheService.flush(workspaceId, ['flatAgentMaps']);
 
         return {
           builtStep: {
@@ -692,33 +676,29 @@ export class WorkflowVersionStepOperationsWorkspaceService {
         };
       }
       case WorkflowActionType.AI_AGENT: {
-        const agentId = step.settings.input.agentId;
+        const existingAgent = await this.agentRepository.findOne({
+          where: { id: step.settings.input.agentId, workspaceId },
+        });
 
-        if (!isDefined(agentId)) {
+        if (!isDefined(existingAgent)) {
           throw new WorkflowVersionStepException(
-            'Agent ID is required for cloning',
+            'Agent not found for cloning',
             WorkflowVersionStepExceptionCode.AI_AGENT_STEP_FAILURE,
           );
         }
 
-        const existingAgent = await this.agentService.findOneAgentById({
-          id: agentId,
+        const clonedAgent = await this.agentRepository.save({
+          name: 'workflow-service-agent' + v4(),
+          label: existingAgent.label,
+          icon: existingAgent.icon,
+          description: existingAgent.description,
+          prompt: existingAgent.prompt,
+          modelId: existingAgent.modelId,
+          responseFormat: existingAgent.responseFormat,
           workspaceId,
+          isCustom: true,
+          modelConfiguration: existingAgent.modelConfiguration,
         });
-
-        const clonedAgent = await this.agentService.createOneAgent(
-          {
-            label: existingAgent.label,
-            icon: existingAgent.icon ?? undefined,
-            description: existingAgent.description ?? undefined,
-            prompt: existingAgent.prompt,
-            modelId: existingAgent.modelId,
-            responseFormat: existingAgent.responseFormat ?? undefined,
-            modelConfiguration: existingAgent.modelConfiguration ?? undefined,
-            isCustom: true,
-          },
-          workspaceId,
-        );
 
         return {
           ...step,
