+++ packages/twenty-server/src/modules/workflow/workflow-tools/tools/update-logic-function-source.tool.ts
@@ -20,7 +20,7 @@ const updateLogicFunctionSourceSchema = z.object({
 });
 
 export const createUpdateLogicFunctionSourceTool = (
-  deps: Pick<WorkflowToolDependencies, 'logicFunctionFromSourceService'>,
+  deps: Pick<WorkflowToolDependencies, 'logicFunctionService'>,
   context: WorkflowToolContext,
 ) => ({
   name: 'update_logic_function_source' as const,
@@ -50,7 +50,7 @@ To find the logicFunctionId, look at the code step's settings.input.logicFunctio
       const { logicFunctionId, code } = parameters;
       const { workspaceId } = context;
 
-      await deps.logicFunctionFromSourceService.updateOne({
+      await deps.logicFunctionService.updateOne({
         id: logicFunctionId,
         update: {
           sourceHandlerCode: code,
