+++ packages/twenty-server/src/modules/workflow/workflow-tools/types/workflow-tool-dependencies.type.ts
@@ -1,5 +1,5 @@
 import type { RecordPositionService } from 'src/engine/core-modules/record-position/services/record-position.service';
-import type { LogicFunctionFromSourceService } from 'src/engine/metadata-modules/logic-function/services/logic-function-from-source.service';
+import type { LogicFunctionService } from 'src/engine/metadata-modules/logic-function/services/logic-function.service';
 import type { WorkspaceManyOrAllFlatEntityMapsCacheService } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.service';
 import type { GlobalWorkspaceOrmManager } from 'src/engine/twenty-orm/global-workspace-datasource/global-workspace-orm.manager';
 import type { WorkflowSchemaWorkspaceService } from 'src/modules/workflow/workflow-builder/workflow-schema/workflow-schema.workspace-service';
@@ -18,7 +18,7 @@ export type WorkflowToolDependencies = {
   workflowSchemaService: WorkflowSchemaWorkspaceService;
   globalWorkspaceOrmManager: GlobalWorkspaceOrmManager;
   recordPositionService: RecordPositionService;
-  logicFunctionFromSourceService: LogicFunctionFromSourceService;
+  logicFunctionService: LogicFunctionService;
   flatEntityMapsCacheService: WorkspaceManyOrAllFlatEntityMapsCacheService;
 };
